requirements.txt
streamlit 
pypdf2 
langchain 
python-dotenv 
faiss-cpu 
openai 
huggingface_hub
langchain_community
tiktoken
InstructorEmbedding
sentence_transformers
